refactor(data-contracts): group logic by feature
