+++ modules/editor/src/main/typescript/data-contract/examples/example-value-provider.ts
