+++ modules/editor/src/main/typescript/data-contract/schema/commands.test.ts
@@ -10,7 +10,7 @@ import {
   executeSchemaCommand,
   findFieldPath,
   updateFieldInTree,
-} from './schemaCommands.js';
+} from './commands.js';
 
 // =============================================================================
 // Test fixtures
