+++ modules/editor/src/main/typescript/data-contract/schema/contract-schema.ts
