+++ modules/editor/src/main/typescript/data-contract/schema/snapshot-history.ts
