+++ modules/editor/src/main/typescript/data-contract/schema/validation.ts
